more changes to migration, changed phonenumber to big integer

+++ db/migrate/20120826034755_change_phone_numberto_integer.rb
@@ -1,7 +1,7 @@
 class ChangePhoneNumbertoInteger < ActiveRecord::Migration
   def change
     rename_column :workorders, :phonenumber, :phonenumber_remove
-    add_column :workorders, :phonenumber, :integer
+    add_column :workorders, :phonenumber, :bigint
     remove_column :workorders, :phonenumber_remove
   end
 end
